Enum musicbrainz::enums::PersonType

pub enum PersonType {
    Person,
    Group,
    Orchestra,
    Choir,
    Character,
    Other,
}

The PersonType enum is used to state whether an artist is a person, a group, or something else.

Variants
Person

Indicates an individual person.

Group

Indicates a group of people that may or may not have a distinctive name.

Orchestra

Indicates an orchestra (a large instrumental ensemble).

Choir

Indicates a choir/chorus (a large vocal ensemble).

Character

Indicates an individual fictional character.

Other

Anything which does not fit into the above categories.